Move NodeSpec to noderepository
